requirements.txt
llama-index
llama-index-llms-openai-like
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
chromadb
pymupdf
streamlit
pydantic
python-dotenv
sentence-transformers